very basic OIS binding working for keyboard button pressed

+++ renderwindow.go
@@ -23,9 +23,8 @@ func (rw *RenderWindow) IsClosed() bool {
 	return gobool(result)
 }
 
-// TODO: c func takes unsigned int
-func (rw *RenderWindow) Handle() int {
-	return C.render_window_get_hwnd(rw.cptr)
+func (rw *RenderWindow) Handle() uint {
+	return uint(C.render_window_get_hwnd(rw.cptr))
 }
 
 func MessagePump() {
